implement jdbc task

+++ src/main/resources/init_db.sql
@@ -1,5 +1,5 @@
 CREATE TABLE book(
   id INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
   title VARCHAR(100),
-  price REAL
+  price DECIMAL(10, 2)
 );
